test(reef): split transport socket fixture
