+++ extensions/reef/src/transport.test.ts
@@ -12,6 +12,7 @@ import {
   createReefWebSocket,
   type WebSocketLike,
 } from "./transport.js";
+import { ControlledSocket } from "./transport.test-helpers.js";
 import type { InboxEntry, ReefKeys, RelayFriend } from "./types.js";
 
 const ts = 1_752_300_000;
@@ -334,35 +335,6 @@ describe("ReefTransportClient response body bounds", () => {
 });
 
 const INBOX_WEBSOCKET_MAX_PAYLOAD_BYTES = 64 * 1024;
-
-class ControlledSocket {
-  closeCalls = 0;
-  private readonly listeners = new Map<string, Array<(event: unknown) => void>>();
-  private closed = false;
-
-  constructor(private readonly closeImmediately = true) {}
-
-  addEventListener(type: string, listener: (event: unknown) => void): void {
-    const listeners = this.listeners.get(type) ?? [];
-    listeners.push(listener);
-    this.listeners.set(type, listeners);
-  }
-
-  close(): void {
-    this.closeCalls++;
-    if (this.closed || !this.closeImmediately) {
-      return;
-    }
-    this.closed = true;
-    this.emit("close");
-  }
-
-  emit(type: string, event: unknown = {}): void {
-    for (const listener of this.listeners.get(type) ?? []) {
-      listener(event);
-    }
-  }
-}
 
 function receiptEntry(seq: number): InboxEntry {
   return {
